Changed location model ViewedStartups
